+++ CHANGES.md
@@ -1,4 +1,4 @@
-# UNRELEASED
+# 2.11.0 (2017-11-21)
 - [NEW] Added an extra bluemix method to the client builder allowing a custom service name to be
   used with the VCAP_SERVICES environment variable content.
 - [FIXED] An issue where `Changes.hasNext()` never returns on receipt of `last_seq` for continuous